TF1 CONV if statement fix
The length of formula is now chached and reused. Furthermore strncmp is used instead of TString for preformance reasons.

+++ core/base/src/TString.cxx
@@ -131,7 +131,7 @@ TString::TString(const char *cs, Ssiz_t n)
       Zero();
       return;
    }
-   if (strlen(cs) < n) {
+   if (strlen(cs) < (size_t)n) {
       Warning("TString::TString", "Input string is shorter than requested size.");
    }
    char *data = Init(n, n);
